Create abstract b2c login component handler
